Improved Performance
Originally, I used arc4random_uniform, because it was the only standard library
function that could provide uniform distribution without a modulo bias. However
the arc4random functions are a little overkill for this, so I sought to use
random, and a custom function to eliminate the modulo bias.

The arc4random version performed 100,000,000 runs in an average 9.5s. The
random based version, did it in 7.986 on average. That's savings of around 16%